Differentiate between State and Event
Events are triggered, completed, and reset, as one action.
LiveData is Lifecycle Aware - does not notify UI of updates onStop(),
and notifies UI of current state onStart(), even if there are no
changes. Thus, if an event is not reset after completion, it persists
on onStart() (e.g. after configuration change).