Display information to user with Snackbar
Snackbars are the preferred way to display feedback messages to users,
as they are displayed in the context where the action occurred. Toasts
persist even when activity is finished, and may be used for system-level
notifications.
Snackbars may perform actions, handle user input, and may be dismissed
by swiping (requires CoordinatorLayout).