Add RecyclerView and Adapter (MVP)
RecyclerView is efficient for displaying large lists, or data that
frequently changes. RecyclerView does minimal work necessary - only
binding data to views displayed on screen at the time, and updating
modified items (instead of the entire list). RecyclerView recycles
views, minimizing space and time (i.e. inflating new views).
Adapter describes how to use the data for RecyclerView. It provides a
method for binding the data onto ViewHolders and a method for creating
ViewHolders, the structure of an item view.