Add Binding Adapters
Binding adapters adapt data for use by views (e.g. specifying how to set
values of the view with provided data). Binding adapters use extension
functions on the view, and the view references the logic in the layout.
Binding adapters separate updating the views from the ViewHolder, and
make it the views responsibility.