Display a header in RecyclerView
An option for displaying differently typed items in RecyclerView is
to modify the list of items to include the different types, and handle
the items based on its type.
The different types extend a sealed class DataItem, and RecyclerView
differentiates types with getItemViewType. RecyclerView needs to
differentiate the types to create appropriate ViewHolders for items
and to ensure only compatible ViewHolders are recycled for an item.